+++ uc-crux-llvm/src/UCCrux/LLVM/Constraints.hs
@@ -83,8 +83,9 @@ import           UCCrux.LLVM.Context.Module (ModuleContext, funcTypes, globalTyp
 import           UCCrux.LLVM.Cursor (Cursor, Selector(..), SomeInSelector(SomeInSelector), seekType, checkCompatibility)
 import           UCCrux.LLVM.Shape (Shape, ShapeSeekError)
 import qualified UCCrux.LLVM.Shape as Shape
-import           UCCrux.LLVM.FullType.Translation (GlobalSymbol, globalSymbol, getGlobalSymbol, FuncSymbol, funcSymbol, getFuncSymbol, ftRetType)
+import           UCCrux.LLVM.FullType.Translation (ftRetType)
 import           UCCrux.LLVM.FullType.Type (FullType(..), FullTypeRepr(FTPtrRepr), ModuleTypes, asFullType)
+import           UCCrux.LLVM.Module (GlobalSymbol, globalSymbol, getGlobalSymbol, FuncSymbol, funcSymbol, getFuncSymbol)
 
 -- See comment in below block of CPP
 #if __GLASGOW_HASKELL__ <= 810
