+++ uc-crux-llvm/src/UCCrux/LLVM/Overrides/Skip.hs
@@ -69,7 +69,8 @@ import           UCCrux.LLVM.Context.Module (ModuleContext, funcTypes, moduleTyp
 import           UCCrux.LLVM.Cursor (Selector(SelectReturn), Cursor(Here))
 import           UCCrux.LLVM.Errors.Panic (panic)
 import           UCCrux.LLVM.FullType.CrucibleType (toCrucibleType)
-import           UCCrux.LLVM.FullType.Translation (FunctionTypes, FuncSymbol, funcSymbol, ftRetType, isDebug, makeFuncSymbol)
+import           UCCrux.LLVM.FullType.Translation (FunctionTypes, ftRetType)
+import           UCCrux.LLVM.Module (FuncSymbol, funcSymbol, makeFuncSymbol, isDebug)
 import           UCCrux.LLVM.Setup (SymValue(getSymValue), generate)
 import           UCCrux.LLVM.Setup.Assume (assume)
 import           UCCrux.LLVM.Setup.Monad (TypedSelector, runSetup, resultAssumptions, resultMem, ppSetupError, resultAnnotations)
