+++ uc-crux-llvm/uc-crux-llvm.cabal
@@ -54,6 +54,7 @@ library
     UCCrux.LLVM.Main
     UCCrux.LLVM.Main.Config.FromEnv
     UCCrux.LLVM.Main.Config.Type
+    UCCrux.LLVM.Module
     UCCrux.LLVM.Newtypes.FunctionName
     UCCrux.LLVM.Newtypes.Seconds
     UCCrux.LLVM.Overrides.Skip
